+++ src/utils/analytics/createDocs.ts
@@ -2,7 +2,7 @@ import { backendURL } from "../../constants";
 
 const axios = require("axios");
 
-export default async function slackhelp() {
+export default async function createDocs() {
   let resp = await axios.get(`${backendURL}/api/analytics/slack/slackhelp`);
   return resp;
 }